visual-retrieval-colpali/pyproject.toml
[project]
name = "visual-retrieval-colpali"
version = "0.1.0"
description = "Visual retrieval with ColPali"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-fasthtml",
    "huggingface-hub",
    "pyvespa",
    "vespacli",
    "torch",
    "vidore-benchmark[interpretability]>=4.0.0,<5.0.0",
    "colpali-engine",
    "einops",
    "shad4fast>=1.2.1",
]
license = "Apache-2.0"

# dev-dependencies
[project.optional-dependencies]
dev = [
    "ruff",
    "python-dotenv",
]
